Silenced a compiler warning
git-svn-id: http://trace2uml.tigris.org/svn/trace2uml/trunk@111 89f8ffb8-d1fc-0310-92ca-d9a8b2f27bcf

+++ Model/Package_LEOIFVGO/classes/class_GQADNYHS/operations/operation_FWJYVFWK/code.cpp
@@ -3,7 +3,7 @@ int eventsCount = dataBase->GetEventsCount();
 eventQueue.clear();
 eventQueue.reserve(dataBase->GetClassCount());
 
-while (thickness.size() < dataBase->GetClassCount())
+while ((int)thickness.size() < dataBase->GetClassCount())
     thickness.push_back(0);
 
 int i;
